Split users request validation tests
Move the interest theme ID request-validation unit tests into a
focused sibling module so the users HTTP test module stays below the
project line limit.

Keep endpoint behaviour tests in the parent module, including the JSON
serialization contract for the request payload.
